[arcane] Déplace la définition de 'MatItem' et 'CellToAllEnvCellConverter' vers 'arcane/core/materials'
