+++ arcane/ceapart/tools/wrapper/materials/ArcaneSwigCeaMaterials.i
@@ -57,7 +57,7 @@ ARCANE_STD_EXHANDLER
 %include arcane/core/materials/IMeshMaterial.h
 %include arcane/core/materials/IMeshEnvironment.h
 %include arcane/core/materials/IMeshMaterialMng.h
-%include arcane/materials/CellToAllEnvCellConverter.h
+%include arcane/core/materials/CellToAllEnvCellConverter.h
 %include MeshMaterialVariable.i
 %exception;
